+++ frontend/src/main.js
@@ -5,6 +5,7 @@ import App from './App.vue'
 
 // PrimeVue imports
 import PrimeVue from 'primevue/config'
+import Aura from '@primeuix/themes/aura'
 import ToastService from 'primevue/toastservice'
 import Toast from 'primevue/toast'
 import 'primeicons/primeicons.css'
@@ -12,7 +13,19 @@ import 'primeicons/primeicons.css'
 const app = createApp(App)
 
 // Use PrimeVue
-app.use(PrimeVue)
+app.use(PrimeVue, {
+    // Default theme configuration
+    theme: {
+        preset: Aura,
+        options: {
+            prefix: 'p',
+            //darkModeSelector: 'system',
+            darkModeSelector: false || 'none',
+            cssLayer: false
+        }
+    }
+});
+
 app.use(ToastService)
 
 // Register global components